Python GUI using the Tkinter module
Python offers various modules to create graphics programs. Out of these Tkinter provides the fastest and easiest way to create GUI applications.
In this mini-project, I made a simple python GUI using the Tkinter module. The Tkinter package ("Tk interface") is the standard Python interface to the Tk GUI toolkit. Both Tk and Tkinter are available on most Unix platforms, as well as on Windows systems.